all: Decouple pkg/rpcmiddleware/discover tests
